+++ src/java/hps/comments/CommentsDAO.java
@@ -88,4 +88,40 @@ public class CommentsDAO implements Serializable{
         }
         return avg;
     }
+    public int getRateStar(String mentorID)
+            throws NamingException, SQLException {
+        Connection con = null;
+        PreparedStatement stm = null;
+        ResultSet rs = null;
+        int rate = 0;
+
+        try {
+            con = DBHelper.makeConnection();
+            if (con != null) {
+                String sql = "Select Ceiling(Avg(rate)) as star "
+                        + "From comments "
+                        + "Where mentorID = ? ";
+                stm = con.prepareStatement(sql);
+                stm.setString(1, mentorID);
+                rs = stm.executeQuery();
+
+                if (rs.next()) {
+                    if (rs.getString("star") != null) {
+                        rate = Integer.parseInt(rs.getString("star"));
+                    }
+                }
+            }
+        } finally {
+            if (rs != null) {
+                rs.close();
+            }
+            if (stm != null) {
+                stm.close();
+            }
+            if (con != null) {
+                con.close();
+            }
+        }
+        return rate;
+    }
 }
