+++ src/java/hps/requests/RequestsDAO.java
@@ -271,7 +271,7 @@ public class RequestsDAO implements Serializable {
                         + "deadline, title, reqContent, status, openedTime, "
                         + "approvedTime, canceledTime, closedTime "
                         + "From requests "
-                        + "Where menteeID like ?";
+                        + "Where menteeID = ?";
                 stm = con.prepareStatement(sql);
                 stm.setString(1, userID);
                 rs = stm.executeQuery();
@@ -350,7 +350,7 @@ public class RequestsDAO implements Serializable {
         try {
             con = DBHelper.makeConnection();
             if (con != null) {
-                String sql = "Delete From requests Where requestID like ?";
+                String sql = "Delete From requests Where requestID = ?";
                 stm = con.prepareStatement(sql);
                 stm.setString(1, requestID);
                 int row = stm.executeUpdate();
@@ -380,7 +380,7 @@ public class RequestsDAO implements Serializable {
                 String sql = "Update requests "
                         + "Set title = ? , deadline = ? , "
                         + "reqContent = ? , skillsID = ? "
-                        + "Where requestID like ?";
+                        + "Where requestID = ?";
                 stm = con.prepareStatement(sql);
                 stm.setString(1, title);
                 stm.setString(2, deadline);
@@ -416,7 +416,7 @@ public class RequestsDAO implements Serializable {
                 String sql = "Update requests "
                         + "Set title = ? , deadline = ? , "
                         + "reqContent = ? , skillsID = ?, mentorID = ? "
-                        + "Where requestID like ?";
+                        + "Where requestID = ?";
                 stm = con.prepareStatement(sql);
                 stm.setString(1, title);
                 stm.setString(2, deadline);
@@ -490,7 +490,7 @@ public class RequestsDAO implements Serializable {
             if (con != null) {
                 String sql = "Select title "
                         + "From requests "
-                        + "Where menteeID like ?";
+                        + "Where menteeID = ?";
                 stm = con.prepareStatement(sql);
                 stm.setString(1, menteeID);
                 rs = stm.executeQuery();
@@ -525,7 +525,7 @@ public class RequestsDAO implements Serializable {
             if (con != null) {
                 String sql = "Select count(*) as totalRequest "
                         + "From requests "
-                        + "Where menteeID like ?";
+                        + "Where menteeID = ?";
                 stm = con.prepareStatement(sql);
                 stm.setString(1, menteeID);
                 rs = stm.executeQuery();
@@ -561,7 +561,7 @@ public class RequestsDAO implements Serializable {
             if (con != null) {
                 String sql = "Select DATEDIFF(second,approvedTime,closedTime)/(60*60) as hours "
                         + "From requests "
-                        + "Where menteeID like ?";
+                        + "Where menteeID = ?";
                 stm = con.prepareStatement(sql);
                 stm.setString(1, menteeID);
                 rs = stm.executeQuery();
@@ -602,7 +602,7 @@ public class RequestsDAO implements Serializable {
             if (con != null) {
                 String sql = "Select COUNT(DISTINCT(mentorID)) as totalMentor "
                         + "From requests "
-                        + "Where menteeID like ?";
+                        + "Where menteeID = ?";
                 stm = con.prepareStatement(sql);
                 stm.setString(1, menteeID);
                 rs = stm.executeQuery();
@@ -641,7 +641,7 @@ public class RequestsDAO implements Serializable {
                         + "deadline, title, reqContent, status, openedTime, "
                         + "approvedTime, canceledTime, closedTime "
                         + "From requests "
-                        + "Where menteeID like ?";
+                        + "Where menteeID = ?";
                 stm = con.prepareStatement(sql);
                 stm.setString(1, menteeID);
 
@@ -689,7 +689,7 @@ public class RequestsDAO implements Serializable {
                         + "deadline, title, reqContent, status, openedTime, "
                         + "approvedTime, canceledTime, closedTime "
                         + "From requests "
-                        + "Where requestID like ?";
+                        + "Where requestID = ?";
                 stm = con.prepareStatement(sql);
                 stm.setString(1, requestID);
 
@@ -768,7 +768,7 @@ public class RequestsDAO implements Serializable {
                 String sql = "Select requestID, title, mentorID "
                         + "From requests "
                         + "Where closedTime is not null AND mentorID is not null "
-                        + "AND menteeID like ? ";
+                        + "AND menteeID = ? ";
                 stm = con.prepareStatement(sql);
                 stm.setString(1, menteeID);
                 rs = stm.executeQuery();
@@ -903,5 +903,74 @@ public class RequestsDAO implements Serializable {
 
         return totalRequest;
     }
-    
+    public int getNumOfApprovedReq(String mentorID)
+            throws NamingException, SQLException {
+        Connection con = null;
+        PreparedStatement stm = null;
+        ResultSet rs = null;
+        int aReq = 0;
+
+        try {
+            con = DBHelper.makeConnection();
+            if (con != null) {
+                String sql = "Select count(*) as numReq "
+                        + "From requests "
+                        + "Where approvedTime is not null AND mentorID = ? ";
+                stm = con.prepareStatement(sql);
+                stm.setString(1, mentorID);
+                rs = stm.executeQuery();
+
+                if (rs.next()) {
+                    aReq = Integer.parseInt(rs.getString("numReq"));
+                }
+            }
+        } finally {
+            if (rs != null) {
+                rs.close();
+            }
+            if (stm != null) {
+                stm.close();
+            }
+            if (con != null) {
+                con.close();
+            }
+        }
+        return aReq;
+    }
+
+    public int getNumOfCompletedReq(String mentorID)
+            throws NamingException, SQLException {
+        Connection con = null;
+        PreparedStatement stm = null;
+        ResultSet rs = null;
+        int cReq = 0;
+
+        try {
+            con = DBHelper.makeConnection();
+            if (con != null) {
+                String sql = "Select count(*) as numReq "
+                        + "From requests "
+                        + "Where approvedTime is not null AND "
+                        + "closedTime is not null AND mentorID = ? ";
+                stm = con.prepareStatement(sql);
+                stm.setString(1, mentorID);
+                rs = stm.executeQuery();
+
+                if (rs.next()) {
+                    cReq = Integer.parseInt(rs.getString("numReq"));
+                }
+            }
+        } finally {
+            if (rs != null) {
+                rs.close();
+            }
+            if (stm != null) {
+                stm.close();
+            }
+            if (con != null) {
+                con.close();
+            }
+        }
+        return cReq;
+    }
 }
